django/requirements.txt
Django==1.9.5
django-filter==0.13.0
djangorestframework==3.3.3
psycopg2==2.6.1
redis==2.10.5
